String analysis improvement, etc.
